+++ src/AutoRest.CSharp.V3/ClientModels/Serialization/XmlObjectElementSerialization.cs
@@ -1,0 +1,22 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace AutoRest.CSharp.V3.ClientModels.Serialization
+{
+    internal class XmlObjectElementSerialization
+    {
+        public XmlObjectElementSerialization(
+            string name,
+            string memberName,
+            XmlSerialization valueSerialization)
+        {
+            Name = name;
+            MemberName = memberName;
+            ValueSerialization = valueSerialization;
+        }
+
+        public string Name { get; }
+        public string MemberName { get; }
+        public XmlSerialization ValueSerialization { get; }
+    }
+}
