+++ src/AutoRest.CSharp.V3/ClientModels/Serialization/XmlSerialization.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace AutoRest.CSharp.V3.ClientModels.Serialization
+{
+    internal abstract class XmlSerialization: ObjectSerialization
+    {
+    }
+}
