+++ src/AutoRest.CSharp.V3/ClientModels/Serialization/XmlValueSerialization.cs
@@ -1,0 +1,17 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace AutoRest.CSharp.V3.ClientModels.Serialization
+{
+    internal class XmlValueSerialization: XmlSerialization
+    {
+        public XmlValueSerialization(ClientTypeReference type, SerializationFormat format)
+        {
+            Type = type;
+            Format = format;
+        }
+
+        public override ClientTypeReference Type { get; }
+        public SerializationFormat Format { get; }
+    }
+}
